back/requirements.txt
Django>=5.0.0
djangorestframework
Pillow
mysqlclient>=2.2.0
python-dotenv
django-cors-headers
requests